feat: Implement Copy Markdown button in template preview
